+++ CHANGELOG.md
@@ -4,7 +4,12 @@ Aribot changes will be documented on this page.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [Unreleased]
+
+### Added
+
+- Added `/confession sync` to fix buttons after bot restart.
+- Added cooldowns to various confession commands.
 
 ## [1.0.0] - 2026-08-11